+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/fargate/validate_02_clean.py
@@ -6,6 +6,7 @@ from polars_utils import utils
 from polars_utils.validation import actions as vl
 from polars_utils.validation.constants import GLOBAL_ACTIONS, GLOBAL_THRESHOLDS
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns
+from utils.column_names.slv_job_role_columns import SLVJobRoleColumns as SLVCols
 
 COMPARE_COLS_TO_IMPORT = [
     IndCqcColumns.location_id,
@@ -45,7 +46,27 @@ def main(
         .row_count_match(
             expected_row_count,
             brief=f"Expects {expected_row_count} rows",
-        ).interrogate()
+        )
+        .col_vals_ge(
+            SLVCols.turnover_rate,
+            0,
+            na_pass=True,
+            brief="turnover_rate is greater than or equal to 0",
+        )
+        .col_vals_ge(
+            SLVCols.starter_rate,
+            0,
+            na_pass=True,
+            brief="starter_rate is greater than or equal to 0",
+        )
+        .col_vals_between(
+            SLVCols.vacancy_rate,
+            0,
+            1,
+            na_pass=True,
+            brief="vacancy_rate is between 0 and 1",
+        )
+        .interrogate()
     )
     vl.write_reports(validation, bucket_name, reports_path)